chore: refactor tests to individual components
