+++ ui/pages/settings-v2/types.ts
@@ -1,0 +1,4 @@
+export type SettingItemConfig = {
+  id: string;
+  component: React.FC;
+};
